% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{checkIfRenvActivated}
\alias{checkIfRenvActivated}
\title{Check if renv is activated}
\usage{
checkIfRenvActivated()
}
\value{
TRUE/FALSE if renv is active / not active
}
\description{
Check if renv is activated
}
