% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{createAddApt}
\alias{createAddApt}
\title{Create the add-apt parts for installing from external repos. This MUST be run first.}
\usage{
createAddApt(deps)
}
\arguments{
\item{deps}{Dependency lookup table (package, dependency, type)}
}
\value{

}
\description{
Create the add-apt parts for installing from external repos. This MUST be run first.
}
