% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{createAptGet}
\alias{createAptGet}
\title{Create the text that does the apt-get commands, this needs to run after addapt}
\usage{
createAptGet(deps)
}
\arguments{
\item{deps}{Dependency lookup table (package, dependency, type)}
}
\value{

}
\description{
Create the text that does the apt-get commands, this needs to run after addapt
}
