% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{createRCMD}
\alias{createRCMD}
\title{Execute the R CMD commands}
\usage{
createRCMD(deps)
}
\arguments{
\item{deps}{Dependency lookup table (package, dependency, type)}
}
\value{

}
\description{
Execute the R CMD commands
}
