% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dep_list}
\alias{dep_list}
\title{Dependency List}
\format{
A tibble with 3 columns and many records:
\describe{
\item{package}{Package Name}
\item{dependency}{Actual code for the dependency}
\item{type}{Type of bash code to install}
}
}
\usage{
dep_list
}
\description{
Dependency List
}
\keyword{datasets}
