% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{extractDependencies}
\alias{extractDependencies}
\title{extractDependencies}
\usage{
extractDependencies(yourpackages)
}
\arguments{
\item{yourpackages}{Vector of packages as passed by the user}
}
\value{

}
\description{
extractDependencies
}
