% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{generateDockerText}
\alias{generateDockerText}
\title{Output the thing to copy / paste into your dockerfile}
\usage{
generateDockerText(yourpackages)
}
\arguments{
\item{yourpackages}{Vector of package names}
}
\value{

}
\description{
Output the thing to copy / paste into your dockerfile
}
