% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_get.R
\name{lockFileDeps}
\alias{lockFileDeps}
\title{Get vector of packages in lock file}
\usage{
lockFileDeps(lockfilepath = NULL, ppa = TRUE)
}
\arguments{
\item{lockfilepath}{Must be the path to a lock file generated by RENV. If NULL it will look for renv.lock in project root.}

\item{ppa}{TRUE / FALSE - Output the ppa or the hardcoded link for external repos}
}
\value{
Printout of text to put in your Docker file
}
\description{
Get vector of packages in lock file
}
