% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{extractDependencies}
\alias{extractDependencies}
\title{extractDependencies}
\usage{
extractDependencies(yourpackages)
}
\arguments{
\item{yourpackages}{list of packages passed by user}

\item{dep_list}{dependency list}
}
\value{

}
\description{
extractDependencies
}
