% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{generateDockerText}
\alias{generateDockerText}
\title{Output the thing to copy / paste into your dockerfile}
\usage{
generateDockerText(yourpackages, ppa = TRUE)
}
\arguments{
\item{yourpackages}{Vector of package names}

\item{ppa}{TRUE / FALSE - Output the ppa or the hardcoded link for external repos}
}
\value{
Console output to copy into your docker file
}
\description{
Output the thing to copy / paste into your dockerfile
}
\examples{
somepackages <- c("gert", "sf", "rkafka", "dplyr")
generateDockerText(somepackages)
generateDockerText(somepackages, ppa = FALSE)
}
