% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDockerFile.R
\name{packageVectorFromlockfile}
\alias{packageVectorFromlockfile}
\title{Get vector of packages in lock file}
\usage{
packageVectorFromlockfile(lockfilepath)
}
\arguments{
\item{lockfilepath}{Must be the path to a lock file generated by RENV}
}
\value{
A vector of package names
}
\description{
Get vector of packages in lock file
}
